llm-counsel/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-counsel"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "aiohttp>=3.9",
    "httpx>=0.27",
    "openai>=1.14",
    "anthropic>=0.21",
    "together>=1.0",
    "sentence-transformers>=2.5",
    "faiss-cpu>=1.8",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "streamlit>=1.32",
    "plotly>=5.19",
    "pandas>=2.2",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-xdist>=3.5",
    "respx>=0.21",
    "mypy>=1.9",
    "ruff>=0.3",
    "pre-commit>=3.6",
    "types-pyyaml",
    "types-requests",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"
markers = [
    "live: marks tests as requiring live API calls",
    "slow: marks tests as slow",
]

[tool.coverage.run]
source = ["api", "core", "models"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 85
show_missing = true

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["tests/", "scripts/"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["api", "core", "models", "dashboard", "evaluation"]
